+++ maplibre/sources.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from enum import Enum
-from typing import Union
+from typing import Optional, Union
 
 from pydantic import ConfigDict, Field, computed_field
 
@@ -35,7 +35,8 @@ class GeoJSONSource(Source):
     Examples:
         >>> from maplibre.sources import GeoJSONSource
 
-        >>> source = GeoJSONSource(data="https://docs.mapbox.com/mapbox-gl-js/assets/earthquakes.geojson")
+        >>> geojson = "https://docs.mapbox.com/mapbox-gl-js/assets/earthquakes.geojson"
+        >>> source = GeoJSONSource(data=geojson)
     """
 
     data: Union[str, dict]
@@ -64,7 +65,8 @@ class RasterTileSource(Source):
 
     Examples:
         >>> from maplibre.sources import RasterTileSource
-        >>> raster_source = RasterTileSource(
+
+        >>> raster_tile_source = RasterTileSource(
         ...     tiles=["https://tile.openstreetmap.org/{z}/{x}/{y}.png"],
         ...     tile_size=256,
         ...     min_zoom=0,
@@ -86,3 +88,40 @@ class RasterTileSource(Source):
     @property
     def type(self) -> str:
         return SourceType.RASTER.value
+
+
+class VectorTileSource(Source):
+    """Vector tile source
+
+    Examples:
+        >>> from maplibre.sources import VectorTileSource
+        >>> from maplibre import LayerType, Layer
+
+        >>> vector_tile_source = VectorTileSource(
+        ...     tiles=["https://demotiles.maplibre.org/tiles/{z}/{x}/{y}.pbf"],
+        ...     min_zoom=0,
+        ...     max_zoom=6,
+        ... )
+
+        >>> layer = Layer(
+        ...     type=LayerType.LINE,
+        ...     id="countries",
+        ...     source=vector_tile_source,
+        ...     source_layer="countries",
+        ...     paint={"fill-color": "lightgreen", "fill-outline-color": "black"},
+        ... )
+    """
+
+    attribution: str = None
+    bounds: tuple = None
+    max_zoom: int = Field(None, serialization_alias="maxzoom")
+    min_zoom: int = Field(None, serialization_alias="minzoom")
+    scheme: str = None
+    tiles: Union[tuple, list] = None
+    url: str = None
+    volatile: bool = None
+
+    @computed_field
+    @property
+    def type(self) -> str:
+        return SourceType.VECTOR.value